UnFix slot on zfs-kmod. It was fine?
This reverts commit e93368d1c00ea39be9ba44130933dd9f667a7ca8.

I was confused because the slots aren't the same between `sys-fs/zfs`
and `sys-fs/zfs-kmod`. The previous commit made them be the same
slot, because that makes sense from a consistency perspective.
But while it would be consistent between ZFS packages, it wouldn't
be consistent with what the Gentoo tree was doing, which was
to give zfs-kmod slot numbers based on its package version,
while `sys-fs/zfs` seems to have some other meaning which
so far isn't clear to me. Either way, it will now be more consistent
with Gentoo's slot numbering, albeit this will create some awfully
long slot numbers on `zfs-kmod`.